Project was splitted into 3 subprojects for better maintenance and implementation isolation
