+++ json-schema-everit/src/test/resources/json/data/basic/message-sample-invalid-missing-sfid.json
