+++ json-schema-everit/src/test/resources/json/data/basic/message-sample-valid-1.json
