+++ json-schema-everit/src/test/resources/json/data/basic/physician-sample-invalid-no-cpso.json
