+++ json-schema-everit/src/test/resources/json/data/basic/physician-sample-valid-with-clinic.json
