+++ json-schema-fge/src/test/resources/json/data/basic/physician-sample-invalid-no-firstname.json
@@ -1,0 +1,4 @@
+{
+	"cpso": "12345",
+	"lastName": "Doe"
+}
