+++ json-schema-first-step/.gitignore
@@ -1,0 +1,1 @@
+/target/
