+++ json-schema-first-step/src/main/java/com/mogikanensoftware/json/schema/validation/service/Response.java
