+++ json-schema-first-step/src/main/java/com/mogikanensoftware/json/schema/validation/service/impl/SimpleRequest.java
