+++ json-schema-first-step/src/main/resources/json/schema/basic/physician.json
