+++ pom.xml
@@ -1,0 +1,17 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+
+	<groupId>com.mogikanensoftware.json</groupId>
+	<artifactId>json-schema-validation-experiments</artifactId>
+	<version>1.0-SNAPSHOT</version>
+	<packaging>pom</packaging>
+	<name>json-schema-validation-experiments</name>
+
+
+	<modules>
+		<module>json-schema-first-step</module>
+		<module>json-schema-everit</module>
+		<module>json-schema-fge</module>
+	</modules>
+</project>
